Fixed sqlite import

+++ config/normalize_data/defs.py
@@ -66,6 +66,7 @@ def export_lob_columns(data_dir, batch, jdbc_url, table, table_columns):
                             "-type=text ",
                             "-file = " + txt_file + " ",
                             "-continueOnError=false ",
+                            "-encoding=UTF8 ",
                             "-clobAsFile=true ",
                             "-blobtype=file ",
                             "-showProgress=10000 ",
@@ -135,6 +136,7 @@ def get_tables(sub_systems_dir, sub_system, jdbc_url, driver_jar):
         text_columns = []
         file_columns = []
         column_defs = table_def.findall('column-def')
+        column_defs[:] = sorted(column_defs, key=lambda elem: int(elem.findtext('dbms-position')))
 
         for column_def in column_defs:
             column_name = column_def.find('column-name')
@@ -142,20 +144,23 @@ def get_tables(sub_systems_dir, sub_system, jdbc_url, driver_jar):
             dbms_data_size = column_def.find('dbms-data-size')
             column_name_fixed = column_name.text
 
+            # TODO: Hvilken av datatypene regnes som hhv blob og clob av sqlwb? -> må sjekke i kildekode?
+            # -> Disse regnes som blob: 2004, -4, -3, -2
+            # Clob'er: -16, -1 -> 2005 og 2011 og?
+            # Test om får eksportert til fil ved å bruke dette: CAST("fldØnsker" AS CLOB)
             if int(java_sql_type.text) in (-4, -3, -2, 2004, 2005, 2011, -16):
                 if int(dbms_data_size.text) > 4000:  # TODO: Sikkert ikke riktig for alle datatyper over
                     jdbc = Jdbc(jdbc_url, '', '', '', 'PUBLIC', driver_jar, 'org.h2.Driver', True, True)
                     length_query = f'''SELECT MAX(LENGTH("{column_name_fixed}")) FROM "{table_name.text}"'''
                     result = run_select(jdbc, length_query)
                     max_length = [x[0] for x in result][0]
-                    if max_length is None:
-                        # TODO: Fjerner tom kolonne fra uttrekk. Legg inn tilsvarende for andre her med count?
-                        continue
-
-                    if max_length > 4000:
-                        file_columns.append(column_name_fixed)
-                        file_name_stem = "'" + table_name.text + "_" + column_name_fixed + "_" + "'"
-                        column_name_fixed = file_name_stem + ' || ROWNUM() AS "' + column_name_fixed + '"'
+                    # TODO: Endre senere her slik at tomme felt ikke skrives til text_columns så fjernes i tsv
+                    # -> Må legge inn 'disposed' på kolonne da og ha sjekk mot det i annen kode så det blir riktg ved opplasting
+                    if max_length is not None:
+                        if max_length > 4000:
+                            file_columns.append(column_name_fixed)
+                            file_name_stem = "'" + table_name.text + "_" + column_name_fixed + "_" + "'"
+                            column_name_fixed = file_name_stem + ' || ROWNUM() AS "' + column_name_fixed + '"'
 
             text_columns.append(column_name_fixed)
 
@@ -265,12 +270,16 @@ def normalize_data(project_dir, bin_dir, class_path, java_path, memory, tmp_dir)
                 for text_file in glob.iglob(data_dir + os.path.sep + '*_lobs.txt'):
                     os.remove(text_file)
 
-        if len(os.listdir(data_docs_dir)) == 0:
-            os.rmdir(data_docs_dir)
-        else:
-            tsv_file = os.path.join(sub_systems_dir, sub_system, 'header', 'data_documents.tsv')
-            if not os.path.isfile(tsv_file):
-                run_tika(tsv_file, data_docs_dir, tika_tmp_dir, java_path)
+        # TODO: Return under for raskere test
+        # return
+
+        if os.path.isdir(data_docs_dir):
+            if len(os.listdir(data_docs_dir)) == 0:
+                os.rmdir(data_docs_dir)
+            else:
+                tsv_file = os.path.join(sub_systems_dir, sub_system, 'header', 'data_documents.tsv')
+                if not os.path.isfile(tsv_file):
+                    run_tika(tsv_file, data_docs_dir, tika_tmp_dir, java_path)
 
         # process files:
         docs_dir = os.path.join(sub_systems_dir, sub_system, 'content', 'documents')
